+++ mayan/apps/mailer/tests/literals.py
@@ -2,6 +2,7 @@ from __future__ import unicode_literals
 
 TEST_BODY_HTML = '<strong>test body</strong>'
 TEST_EMAIL_ADDRESS = 'test@example.com'
+TEST_EMAIL_FROM_ADDRESS = 'from.test@example.com'
 TEST_RECIPIENTS_MULTIPLE_COMMA = 'test@example.com,test2@example.com'
 TEST_RECIPIENTS_MULTIPLE_COMMA_RESULT = [
     'test@example.com', 'test2@example.com'