+++ mayan/apps/mailer/tests/mixins.py
@@ -1,8 +1,13 @@
 from __future__ import unicode_literals
 
+import json
+
 from ..models import UserMailer
 
-from .literals import TEST_USER_MAILER_BACKEND_PATH, TEST_USER_MAILER_LABEL
+from .literals import (
+    TEST_EMAIL_FROM_ADDRESS, TEST_USER_MAILER_BACKEND_PATH,
+    TEST_USER_MAILER_LABEL
+)
 
 
 class MailerTestMixin(object):
@@ -12,5 +17,9 @@ class MailerTestMixin(object):
             enabled=True,
             label=TEST_USER_MAILER_LABEL,
             backend_path=TEST_USER_MAILER_BACKEND_PATH,
-            backend_data='{}'
+            backend_data=json.dumps(
+                {
+                    'from': TEST_EMAIL_FROM_ADDRESS
+                }
+            )
         )